+++ Joker.Kafka/ChangeLog.md
@@ -6,10 +6,11 @@
 - Avg - Return the average value for a given column
 - Min
 - Max
-**TODO:**
-- Trim, Len, LPad, RPad, Sign, Sqrt
-- LEN(FIRST_NAME) + LEN(LAST_NAME) AS NAME_LENGTH
 - KSqlFunctions - LIKE
+- String functions - Len
+**TODO:**
+- Trim, LPad, RPad, Sign, Sqrt
+- LEN(FIRST_NAME) + LEN(LAST_NAME) AS NAME_LENGTH
 
 # v0.1.0
 ### ExtensionsMethods: